Move new enum to it's own correct file and location
