+++ Octokit/Clients/OrganizationMembersClient.cs
@@ -37,20 +37,6 @@ namespace Octokit
         Member
     }
 
-    public enum OrganizationMembershipRole
-    {
-        [Parameter(Value = "direct_member")]
-        DirectMember,
-        [Parameter(Value = "admin")]
-        Admin,
-        [Parameter(Value = "billing_manager")]
-        BillingManager,
-        [Parameter(Value = "hiring_manager")]
-        HiringManager,
-        [Parameter(Value = "reinstate")]
-        Reinstate
-    }
-
     /// <summary>
     /// A client for GitHub's Organization Members API.
     /// </summary>
